+++ src/net_client.c
@@ -1,7 +1,7 @@
 // Emacs style mode select   -*- C++ -*- 
 //-----------------------------------------------------------------------------
 //
-// $Id: net_client.c 593 2006-09-01 20:45:45Z fraggle $
+// $Id: net_client.c 641 2006-09-21 11:13:28Z rtc_marine $
 //
 // Copyright(C) 2005 Simon Howard
 //
@@ -596,7 +596,7 @@ static void NET_CL_ParseWaitingData(net_packet_t *packet)
     unsigned int player_number;
     char *player_names[MAXPLAYERS];
     char *player_addr[MAXPLAYERS];
-    int i;
+    size_t i;
 
     if (!NET_ReadInt8(packet, &num_players)
      || !NET_ReadInt8(packet, &is_controller)
@@ -804,7 +804,7 @@ static void NET_CL_ParseGameData(net_packet_t *packet)
     unsigned int seq, num_tics;
     unsigned int nowtime;
     int resend_start, resend_end;
-    int i;
+    size_t i;
     int index;
     
     // Read header
